+++ vitrivr-engine-query/src/main/kotlin/org/vitrivr/engine/query/operators/transform/scoring/ScoreAggregator.kt
@@ -1,4 +1,4 @@
-package org.vitrivr.engine.query.transform
+package org.vitrivr.engine.query.operators.transform.scoring
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
